+++ CadQuery/Libs/cadquery-lib/tests/TestCQSelectors.py
@@ -166,7 +166,51 @@ class TestCQSelectors(BaseTest):
         # test the case of multiple objects at the same distance
         el = c.edges("<Z").vals()
         self.assertEqual(4, len(el))
-
+        
+    def testNthDistance(self):
+        c = Workplane('XY').pushPoints([(-2,0),(2,0)]).box(1,1,1)
+        
+        #2nd face
+        val = c.faces(selectors.DirectionNthSelector(Vector(1,0,0),1)).val()
+        self.assertAlmostEqual(val.Center().x,-1.5)
+        
+        #2nd face with inversed selection vector
+        val = c.faces(selectors.DirectionNthSelector(Vector(-1,0,0),1)).val()
+        self.assertAlmostEqual(val.Center().x,1.5)
+        
+        #2nd last face
+        val = c.faces(selectors.DirectionNthSelector(Vector(1,0,0),-2)).val()
+        self.assertAlmostEqual(val.Center().x,1.5)
+        
+        #Last face
+        val = c.faces(selectors.DirectionNthSelector(Vector(1,0,0),-1)).val()
+        self.assertAlmostEqual(val.Center().x,2.5)
+        
+        #check if the selected face if normal to the specified Vector
+        self.assertAlmostEqual(val.normalAt().cross(Vector(1,0,0)).Length,0.0)
+        
+        #repeat the test using string based selector
+        
+        #2nd face
+        val = c.faces('>(1,0,0)[1]').val()
+        self.assertAlmostEqual(val.Center().x,-1.5)
+        
+        #2nd face with inversed selection vector
+        val = c.faces('>(-1,0,0)[1]').val()
+        self.assertAlmostEqual(val.Center().x,1.5)
+        
+        #2nd last face
+        val = c.faces('>X[-2]').val()
+        self.assertAlmostEqual(val.Center().x,1.5)
+        
+        #Last face
+        val = c.faces('>X[-1]').val()
+        self.assertAlmostEqual(val.Center().x,2.5)
+        
+        #check if the selected face if normal to the specified Vector
+        self.assertAlmostEqual(val.normalAt().cross(Vector(1,0,0)).Length,0.0)
+        
+        
     def testNearestTo(self):
         c = CQ(makeUnitCube())
 
@@ -356,3 +400,30 @@ class TestCQSelectors(BaseTest):
         #make sure the vertex is the right one
 
         self.assertTupleAlmostEquals((0.0,0.0,1.0),v2.val().toTuple() ,3)
+        
+    def testGrammar(self):
+        """
+        Test if reasonable string selector expressions parse without an error
+        """
+        
+        gram = selectors._makeGrammar()
+
+        expressions = ['+X ',
+                       '-Y',
+                       '|(1,0,0)',
+                       '#(1.,1.4114,-0.532)',
+                       '%Plane',
+                       '>XZ',
+                       '<Z[-2]',
+                       '>(1,4,55.)[20]',
+                       '|XY',
+                       '<YZ[0]',
+                       'front',
+                       'back',
+                       'left',
+                       'right',
+                       'top',
+                       'bottom']
+        
+        for e in expressions: gram.parseString(e)
+